feat: refactor on sqlite setup
