+++ src/app/page.tsx
@@ -1,12 +1,15 @@
 import React, { Suspense } from "react"
-import { add, addDB } from "../components/actions"
+import { update } from "../components/actions"
 import Counter from "../components/counter"
-import CounterDB from "../components/counter-db"
 import OS from "../components/os"
-import { selectQuery } from "../database"
+import Toaster from "../components/toaster"
+import { db } from "../database"
 
 const Page = () => {
-  const { counter } = selectQuery.get() as { counter: number }
+  const { counter } = db.query(`SELECT value as counter FROM counter WHERE id = 1`).get() as {
+    counter: number
+  }
+
   return (
     <>
       <main className="m-4 border-4 border-dashed border-red-400 p-4">
@@ -20,13 +23,11 @@ const Page = () => {
 
         <section className="mt-4 flex h-16 items-center justify-center border-4 border-dashed border-blue-400">
           <Suspense fallback={"Loading counter..."}>
-            <Counter action={add} />
+            <Counter initValue={counter} action={update} />
           </Suspense>
-        </section>
 
-        <section className="mt-4 flex h-16 items-center justify-center border-4 border-dashed border-blue-400">
-          <Suspense fallback={"Loading counter..."}>
-            <CounterDB initValue={counter} action={addDB} />
+          <Suspense fallback={"Loading toaster..."}>
+            <Toaster />
           </Suspense>
         </section>
