+++ src/database.ts
@@ -1,8 +1,8 @@
-import { join, resolve } from "path"
 import { Database } from "bun:sqlite"
 
-export const db = new Database(join(resolve("data"), "db.sqlite"))
+// Create a database instance
+export const db = new Database("db.sqlite")
+
+// Setup a mock database
 db.exec(`CREATE TABLE IF NOT EXISTS counter (id INTEGER PRIMARY KEY, value INTEGER)`)
 db.exec(`INSERT OR IGNORE INTO counter (id, value) VALUES (1, 0)`)
-export const updateQuery = db.query(`UPDATE counter SET value = value + 1 WHERE id = 1`)
-export const selectQuery = db.query(`SELECT value as counter FROM counter WHERE id = 1`)
